create new folder name redux and move store.js to it
