+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import { Provider } from 'react-redux';
 import App from './App';
-import store from './app/store';
+import store from './redux/store';
 
 const domContainer = document.getElementById('root');
 const root = ReactDOM.createRoot(domContainer);
